refactor: WorlModalからWorkAssetを分離
